refactoring and cleanup
