+++ src/components/Backlog.js
@@ -10,7 +10,6 @@ import { useTheme } from '@mui/styles';
 
 import { useAppContext } from '../AppContext';
 
-import { counter } from '../lib/functions';
 import { addGame, getGames, deleteGame, updateGame } from '../lib/games';
 import {
 	AddGameModal,
@@ -20,12 +19,6 @@ import {
 	GameCards,
 	LoadingOverlay,
 } from './components';
-
-const handleSorting = (a, b) => {
-	var titleA = a.title.toUpperCase();
-	var titleB = b.title.toUpperCase();
-	return titleA < titleB ? -1 : titleA > titleB ? 1 : 0;
-};
 
 const Backlog = () => {
 	const { data, status } = useSession();
@@ -40,7 +33,6 @@ const Backlog = () => {
 	const [filteredGames, setFilteredGames] = useState();
 	const [openModal, setOpenModal] = useState(false);
 	const [viewMode, setViewMode] = useState('grid');
-	const [statusCounter, setStatusCounter] = useState({});
 
 	useEffect(() => {
 		if (games) {
@@ -101,13 +93,6 @@ const Backlog = () => {
 		Router.push(`/backlog?tab=${newStatus}`);
 	};
 
-	useEffect(() => {
-		if (games) {
-			const statusCounts = counter(games);
-			setStatusCounter(statusCounts);
-		}
-	}, [games]);
-
 	return (
 		<>
 			{!status === 'authenticated' ? null : (
@@ -153,7 +138,7 @@ const Backlog = () => {
 							</Fab>
 						</Box>
 						<GameCards
-							games={filteredGames?.sort((a, b) => handleSorting(a, b))}
+							games={filteredGames}
 							loading={loading}
 							setStatus={setStatus}
 						/>
